requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
PyJWT==2.4.0